requirements.txt
Django~=5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
celery==5.4.0
django-celery-beat==2.7.0
python-dotenv==1.0.1